refactor: pull walks logic out of writer into separate util
